save0rUpdate product

+++ fullstack-ecommerce-main/src/backend/src/main/java/org/stand/springbootecommerce/entity/Flavours.java
@@ -11,7 +11,7 @@ public class Flavours  {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
     @Column(name = "name")
     private String name;
     @Column(name = "description")
@@ -20,15 +20,15 @@ public class Flavours  {
     public Flavours() {
     }
 
-    public Flavours(Integer id) {
+    public Flavours(Long id) {
         this.id = id;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
